requirements.txt
3==0.0.0
aiofiles==0.8.0
pyshorteners==1.0.1
aiogram==3.0.0b2
aiogram-dialog==2.0.0b3
aiohttp==3.8.1
aiohttp-jinja2==1.5
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
ciso8601==2.2.0
Deprecated==1.2.13
diagrams
environs==9.5.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
influxdb-client==1.27.0
install==1.3.5
Jinja2==3.0.3
magic-filter==1.0.5
MarkupSafe==2.1.0
marshmallow==3.14.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
pandas==1.4.1
psycopg2-binary==2.9.3
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.2.2
Rx==3.2.0
six==1.16.0
slugify==0.0.1
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
yarl==1.7.2
